+++ packages/metadata/data/layers/full/tokens/js/index.d.ts
@@ -1,5 +1,5 @@
 /**
- * @synergy-design-system/tokens version 3.18.0
+ * @synergy-design-system/tokens version 3.20.2
  * SICK Global UX Foundation
  */